feat: show Settings, Tasks, and Cronjobs as popup dialogs on web
Instead of navigating to separate routes (which replaces the main view),
these pages now open as overlay dialogs on desktop web. The main
dashboard stays visible behind them. Tauri behavior unchanged (native
windows). Direct URL routes still work as standalone full-page views.

- Extract TasksPageContent and CronjobsPageContent from route files
- Add dialog wrappers in ToolbarButtons and DashboardShell
- Settings always opens as dialog on web (remove navigate path)
- Standalone routes bypass desktop sidebar layout
- Settings detail panel scrollable within fixed-height dialog
