+++ apps/web/src/routeTree.gen.ts
@@ -10,6 +10,7 @@
 
 import { Route as rootRouteImport } from './routes/__root'
 import { Route as TasksRouteImport } from './routes/tasks'
+import { Route as SettingsRouteImport } from './routes/settings'
 import { Route as CronjobsRouteImport } from './routes/cronjobs'
 import { Route as IndexRouteImport } from './routes/index'
 import { Route as WorkspaceWorkspaceIdRouteImport } from './routes/workspace.$workspaceId'
@@ -23,6 +24,11 @@ import { Route as WorkspaceWorkspaceIdCodeSplatRouteImport } from './routes/work
 const TasksRoute = TasksRouteImport.update({
   id: '/tasks',
   path: '/tasks',
+  getParentRoute: () => rootRouteImport,
+} as any)
+const SettingsRoute = SettingsRouteImport.update({
+  id: '/settings',
+  path: '/settings',
   getParentRoute: () => rootRouteImport,
 } as any)
 const CronjobsRoute = CronjobsRouteImport.update({
@@ -80,6 +86,7 @@ const WorkspaceWorkspaceIdCodeSplatRoute =
 export interface FileRoutesByFullPath {
   '/': typeof IndexRoute
   '/cronjobs': typeof CronjobsRoute
+  '/settings': typeof SettingsRoute
   '/tasks': typeof TasksRoute
   '/workspace/$workspaceId': typeof WorkspaceWorkspaceIdRouteWithChildren
   '/workspace/$workspaceId/changes': typeof WorkspaceWorkspaceIdChangesRoute
@@ -92,6 +99,7 @@ export interface FileRoutesByFullPath {
 export interface FileRoutesByTo {
   '/': typeof IndexRoute
   '/cronjobs': typeof CronjobsRoute
+  '/settings': typeof SettingsRoute
   '/tasks': typeof TasksRoute
   '/workspace/$workspaceId/changes': typeof WorkspaceWorkspaceIdChangesRoute
   '/workspace/$workspaceId/terminal': typeof WorkspaceWorkspaceIdTerminalRoute
@@ -103,6 +111,7 @@ export interface FileRoutesById {
   __root__: typeof rootRouteImport
   '/': typeof IndexRoute
   '/cronjobs': typeof CronjobsRoute
+  '/settings': typeof SettingsRoute
   '/tasks': typeof TasksRoute
   '/workspace/$workspaceId': typeof WorkspaceWorkspaceIdRouteWithChildren
   '/workspace/$workspaceId/changes': typeof WorkspaceWorkspaceIdChangesRoute
@@ -117,6 +126,7 @@ export interface FileRouteTypes {
   fullPaths:
     | '/'
     | '/cronjobs'
+    | '/settings'
     | '/tasks'
     | '/workspace/$workspaceId'
     | '/workspace/$workspaceId/changes'
@@ -129,6 +139,7 @@ export interface FileRouteTypes {
   to:
     | '/'
     | '/cronjobs'
+    | '/settings'
     | '/tasks'
     | '/workspace/$workspaceId/changes'
     | '/workspace/$workspaceId/terminal'
@@ -139,6 +150,7 @@ export interface FileRouteTypes {
     | '__root__'
     | '/'
     | '/cronjobs'
+    | '/settings'
     | '/tasks'
     | '/workspace/$workspaceId'
     | '/workspace/$workspaceId/changes'
@@ -152,6 +164,7 @@ export interface FileRouteTypes {
 export interface RootRouteChildren {
   IndexRoute: typeof IndexRoute
   CronjobsRoute: typeof CronjobsRoute
+  SettingsRoute: typeof SettingsRoute
   TasksRoute: typeof TasksRoute
   WorkspaceWorkspaceIdRoute: typeof WorkspaceWorkspaceIdRouteWithChildren
 }
@@ -163,6 +176,13 @@ declare module '@tanstack/react-router' {
       path: '/tasks'
       fullPath: '/tasks'
       preLoaderRoute: typeof TasksRouteImport
+      parentRoute: typeof rootRouteImport
+    }
+    '/settings': {
+      id: '/settings'
+      path: '/settings'
+      fullPath: '/settings'
+      preLoaderRoute: typeof SettingsRouteImport
       parentRoute: typeof rootRouteImport
     }
     '/cronjobs': {
@@ -267,6 +287,7 @@ const WorkspaceWorkspaceIdRouteWithChildren =
 const rootRouteChildren: RootRouteChildren = {
   IndexRoute: IndexRoute,
   CronjobsRoute: CronjobsRoute,
+  SettingsRoute: SettingsRoute,
   TasksRoute: TasksRoute,
   WorkspaceWorkspaceIdRoute: WorkspaceWorkspaceIdRouteWithChildren,
 }
