+++ packages/dashboard-core/src/components/SettingsPage.tsx
@@ -60,39 +60,57 @@ type Section =
   | "web-server"
   | "labels";
 
+const SECTION_TITLES: Record<Exclude<Section, "menu">, string> = {
+  appearance: "Appearance",
+  general: "General",
+  labels: "Labels",
+  "coding-agent": "Coding Agent",
+  defaults: "Workspace Settings",
+  notifications: "Notifications",
+  "web-server": "Web Server",
+};
+
 interface Props {
-  onClose: () => void;
+  onClose?: () => void;
+  hideTitle?: boolean;
 }
 
 function SettingsRow({
   label,
   value,
+  active,
   onClick,
 }: {
   label: string;
   value?: string;
+  active?: boolean;
   onClick: () => void;
 }) {
   return (
     <button
       type="button"
-      className="flex w-full items-center justify-between px-3 py-2.5 text-sm hover:bg-accent/50 rounded-md transition-colors text-left"
+      className={`flex w-full items-center justify-between px-3 py-2.5 text-sm hover:bg-accent/50 rounded-md transition-colors text-left ${active ? "bg-accent/50" : ""}`}
       onClick={onClick}
     >
       <span>{label}</span>
       <span className="flex items-center gap-1 text-muted-foreground">
         {value && <span className="text-xs truncate max-w-[140px]">{value}</span>}
-        <ChevronRight className="size-4 shrink-0" />
+        <ChevronRight className="size-4 shrink-0 lg:hidden" />
       </span>
     </button>
   );
 }
 
-export function SettingsPage({ onClose }: Props) {
+export function SettingsPage({ onClose, hideTitle }: Props) {
   const { settings } = useSettingsQuery();
   const updateSettingsMutation = useUpdateSettings();
   const capabilities = useCapabilities();
-  const [section, setSection] = useState<Section>("menu");
+  const [section, setSection] = useState<Section>(() => {
+    if (typeof window !== "undefined" && window.matchMedia("(min-width: 1024px)").matches) {
+      return "appearance";
+    }
+    return "menu";
+  });
   const [worktreesDir, setWorktreesDir] = useState(settings.worktreesDir ?? "");
   const [defaultsJson, setDefaultsJson] = useState("");
   const [defaultsError, setDefaultsError] = useState<string | null>(null);
@@ -238,159 +256,246 @@ export function SettingsPage({ onClose }: Props) {
     ? (SOUNDS.find((s) => s.id === selectedSound)?.label ?? "On")
     : "Off";
 
-  if (section !== "menu") {
-    return (
-      <div>
-        <div className="flex items-center gap-1 mb-3 px-1">
-          <Button variant="ghost" size="icon-sm" onClick={() => setSection("menu")}>
-            <ChevronLeft className="size-5" />
-          </Button>
-          <h2 className="text-base font-semibold flex-1">
-            {section === "appearance" && "Appearance"}
-            {section === "general" && "General"}
-            {section === "labels" && "Labels"}
-            {section === "coding-agent" && "Coding Agent"}
-            {section === "defaults" && "Workspace Settings"}
-            {section === "notifications" && "Notifications"}
-            {section === "web-server" && "Web Server"}
-          </h2>
-          <Tooltip>
-            <TooltipTrigger asChild>
+  const activeSection = section === "menu" ? null : section;
+
+  /* ── Shared section content ─────────────────────────────── */
+
+  const sectionContent = activeSection && (
+    <>
+      {activeSection === "appearance" && (
+        <div className="space-y-4 px-1">
+          <div className="space-y-2">
+            <Label>Theme</Label>
+            <DropdownMenu>
+              <DropdownMenuTrigger asChild>
+                <Button
+                  variant="outline"
+                  size="sm"
+                  className="w-full justify-between font-normal h-7 text-xs px-2"
+                >
+                  {selectedTheme === "system"
+                    ? "System"
+                    : selectedTheme === "dark"
+                      ? "Dark"
+                      : "Light"}
+                  <ChevronDown className="size-3 opacity-50" />
+                </Button>
+              </DropdownMenuTrigger>
+              <DropdownMenuContent
+                align="start"
+                className="w-[--radix-dropdown-menu-trigger-width]"
+              >
+                <DropdownMenuRadioGroup
+                  value={selectedTheme}
+                  onValueChange={(v: string) => setSelectedTheme(v as Theme)}
+                >
+                  <DropdownMenuRadioItem value="system">System</DropdownMenuRadioItem>
+                  <DropdownMenuRadioItem value="light">Light</DropdownMenuRadioItem>
+                  <DropdownMenuRadioItem value="dark">Dark</DropdownMenuRadioItem>
+                </DropdownMenuRadioGroup>
+              </DropdownMenuContent>
+            </DropdownMenu>
+            <p className="text-xs text-muted-foreground">
+              Choose between system default, light, and dark mode. System follows your OS preference.
+              You can also cycle through themes using the toolbar button.
+            </p>
+          </div>
+        </div>
+      )}
+
+      {activeSection === "general" && (
+        <div className="space-y-4 px-1">
+          <div className="space-y-2">
+            <Label htmlFor="worktrees-dir">Worktrees folder</Label>
+            <div className="flex gap-2">
+              <Input
+                id="worktrees-dir"
+                placeholder="~/.band/worktrees (default)"
+                value={worktreesDir}
+                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+                  setWorktreesDir(e.target.value)
+                }
+              />
+              {capabilities.pickFolder && (
+                <Button type="button" variant="ghost" size="icon" onClick={handleBrowse}>
+                  <FolderOpen />
+                </Button>
+              )}
+            </div>
+            <p className="text-xs text-muted-foreground">
+              Directory where new worktrees are created. Leave empty for the default location.
+            </p>
+          </div>
+        </div>
+      )}
+
+      {activeSection === "labels" && (
+        <div className="space-y-3 px-1">
+          {labels.map((lbl) => (
+            <div key={lbl.id} className="flex items-center gap-2">
+              <ColorPicker
+                value={lbl.color}
+                onChange={(color) =>
+                  setLabels((prev) => prev.map((l) => (l.id === lbl.id ? { ...l, color } : l)))
+                }
+                showHex={false}
+                className="w-auto h-7 px-1.5 shrink-0"
+              />
+              <Input
+                value={lbl.name}
+                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+                  setLabels((prev) =>
+                    prev.map((l) => (l.id === lbl.id ? { ...l, name: e.target.value } : l)),
+                  )
+                }
+                className="flex-1 h-7 text-xs"
+              />
               <Button
                 variant="ghost"
-                size="icon-sm"
-                onClick={handleSave}
-                disabled={section === "defaults" && !!defaultsError}
-                className="relative"
+                size="icon-xs"
+                className="text-destructive hover:text-destructive shrink-0"
+                onClick={() => setLabels((prev) => prev.filter((l) => l.id !== lbl.id))}
               >
-                <Save className="size-5" />
-                {isDirty && (
-                  <span className="absolute top-0.5 right-0.5 size-2 rounded-full bg-blue-500" />
-                )}
+                <Trash2 className="size-3" />
               </Button>
-            </TooltipTrigger>
-            <TooltipContent>Save</TooltipContent>
-          </Tooltip>
+            </div>
+          ))}
+          <Button
+            variant="outline"
+            size="sm"
+            className="w-full text-xs"
+            onClick={() => {
+              const id = `lbl_${Date.now()}`;
+              setLabels((prev) => [...prev, { id, name: "New label", color: "#3b82f6" }]);
+            }}
+          >
+            <Plus className="size-3 mr-1" />
+            Add label
+          </Button>
         </div>
-        <Separator className="mb-3" />
+      )}
 
-        {section === "appearance" && (
-          <div className="space-y-4 px-1">
-            <div className="space-y-2">
-              <Label>Theme</Label>
-              <DropdownMenu>
-                <DropdownMenuTrigger asChild>
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    className="w-full justify-between font-normal h-7 text-xs px-2"
-                  >
-                    {selectedTheme === "system"
-                      ? "System"
-                      : selectedTheme === "dark"
-                        ? "Dark"
-                        : "Light"}
-                    <ChevronDown className="size-3 opacity-50" />
-                  </Button>
-                </DropdownMenuTrigger>
-                <DropdownMenuContent
-                  align="start"
-                  className="w-[--radix-dropdown-menu-trigger-width]"
-                >
-                  <DropdownMenuRadioGroup
-                    value={selectedTheme}
-                    onValueChange={(v: string) => setSelectedTheme(v as Theme)}
-                  >
-                    <DropdownMenuRadioItem value="system">System</DropdownMenuRadioItem>
-                    <DropdownMenuRadioItem value="light">Light</DropdownMenuRadioItem>
-                    <DropdownMenuRadioItem value="dark">Dark</DropdownMenuRadioItem>
-                  </DropdownMenuRadioGroup>
-                </DropdownMenuContent>
-              </DropdownMenu>
-              <p className="text-xs text-muted-foreground">
-                Choose between system default, light, and dark mode. System follows your OS
-                preference. You can also cycle through themes using the toolbar button.
-              </p>
-            </div>
-          </div>
-        )}
-
-        {section === "general" && (
-          <div className="space-y-4 px-1">
-            <div className="space-y-2">
-              <Label htmlFor="worktrees-dir">Worktrees folder</Label>
-              <div className="flex gap-2">
-                <Input
-                  id="worktrees-dir"
-                  placeholder="~/.band/worktrees (default)"
-                  value={worktreesDir}
-                  onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                    setWorktreesDir(e.target.value)
-                  }
-                />
-                {capabilities.pickFolder && (
-                  <Button type="button" variant="ghost" size="icon" onClick={handleBrowse}>
-                    <FolderOpen />
-                  </Button>
-                )}
-              </div>
-              <p className="text-xs text-muted-foreground">
-                Directory where new worktrees are created. Leave empty for the default location.
-              </p>
-            </div>
-          </div>
-        )}
-
-        {section === "labels" && (
-          <div className="space-y-3 px-1">
-            {labels.map((lbl) => (
-              <div key={lbl.id} className="flex items-center gap-2">
-                <ColorPicker
-                  value={lbl.color}
-                  onChange={(color) =>
-                    setLabels((prev) => prev.map((l) => (l.id === lbl.id ? { ...l, color } : l)))
-                  }
-                  showHex={false}
-                  className="w-auto h-7 px-1.5 shrink-0"
-                />
-                <Input
-                  value={lbl.name}
-                  onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                    setLabels((prev) =>
-                      prev.map((l) => (l.id === lbl.id ? { ...l, name: e.target.value } : l)),
-                    )
-                  }
-                  className="flex-1 h-7 text-xs"
-                />
+      {activeSection === "coding-agent" && (
+        <div className="space-y-4 px-1">
+          <div className="space-y-2">
+            <Label>Agent type</Label>
+            <DropdownMenu>
+              <DropdownMenuTrigger asChild>
                 <Button
-                  variant="ghost"
-                  size="icon-xs"
-                  className="text-destructive hover:text-destructive shrink-0"
-                  onClick={() => setLabels((prev) => prev.filter((l) => l.id !== lbl.id))}
+                  variant="outline"
+                  size="sm"
+                  className="w-full justify-between font-normal h-7 text-xs px-2"
                 >
-                  <Trash2 className="size-3" />
+                  {agentType ? AGENT_LABEL[agentType] : "None"}
+                  <ChevronDown className="size-3 opacity-50" />
                 </Button>
-              </div>
-            ))}
-            <Button
-              variant="outline"
-              size="sm"
-              className="w-full text-xs"
-              onClick={() => {
-                const id = `lbl_${Date.now()}`;
-                setLabels((prev) => [...prev, { id, name: "New label", color: "#3b82f6" }]);
-              }}
-            >
-              <Plus className="size-3 mr-1" />
-              Add label
-            </Button>
+              </DropdownMenuTrigger>
+              <DropdownMenuContent
+                align="start"
+                className="w-[--radix-dropdown-menu-trigger-width]"
+              >
+                <DropdownMenuRadioGroup
+                  value={agentType}
+                  onValueChange={(v: string) => setAgentType(v as CodingAgentType | "")}
+                >
+                  <DropdownMenuRadioItem value="">None</DropdownMenuRadioItem>
+                  {AGENT_TYPES.map((t) => (
+                    <DropdownMenuRadioItem key={t.value} value={t.value}>
+                      {t.label}
+                    </DropdownMenuRadioItem>
+                  ))}
+                </DropdownMenuRadioGroup>
+              </DropdownMenuContent>
+            </DropdownMenu>
+            <p className="text-xs text-muted-foreground">
+              The coding agent used for background jobs.
+            </p>
           </div>
-        )}
-
-        {section === "coding-agent" && (
-          <div className="space-y-4 px-1">
+          {agentType && (
             <div className="space-y-2">
-              <Label>Agent type</Label>
+              <Label htmlFor="agent-command">
+                Command <span className="text-muted-foreground font-normal">(optional)</span>
+              </Label>
+              <Input
+                id="agent-command"
+                placeholder="claude --dangerously-skip-permissions"
+                value={agentCommand}
+                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+                  setAgentCommand(e.target.value)
+                }
+              />
+              <p className="text-xs text-muted-foreground">
+                Custom command with arguments to run the agent. Leave empty to use the default
+                command for the selected agent type.
+              </p>
+            </div>
+          )}
+        </div>
+      )}
+
+      {activeSection === "defaults" && (
+        <div className="space-y-4 px-1">
+          <div className="space-y-2">
+            <div className="flex items-center justify-between">
+              <Label htmlFor="defaults-json">Default apps configuration</Label>
+              {!defaultsJson.trim() && (
+                <Button
+                  type="button"
+                  variant="link"
+                  size="sm"
+                  className="h-auto p-0 text-xs"
+                  onClick={handleInsertTemplate}
+                >
+                  Insert template
+                </Button>
+              )}
+            </div>
+            <textarea
+              id="defaults-json"
+              className="w-full min-h-[400px] rounded-md border border-input bg-background px-2 py-1.5 text-xs font-mono ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 dark:bg-input/30"
+              placeholder='{"apps": [{"type": "vscode", ...}]}'
+              value={defaultsJson}
+              onChange={(e: React.ChangeEvent<HTMLTextAreaElement>) =>
+                handleDefaultsChange(e.target.value)
+              }
+              autoCapitalize="off"
+              autoCorrect="off"
+              autoComplete="off"
+              spellCheck={false}
+            />
+            {defaultsError && <p className="text-xs text-destructive">{defaultsError}</p>}
+            <p className="text-xs text-muted-foreground">
+              Default apps configuration applied to Band worktrees that don't have a project-level{" "}
+              <code className="text-xs">.band/config.json</code>. Defines which apps (VS Code,
+              Cursor, Zed, iTerm, Chrome) to open and their layout. Leave empty to disable.
+            </p>
+          </div>
+        </div>
+      )}
+
+      {activeSection === "notifications" && (
+        <div className="space-y-4 px-1">
+          <div className="space-y-2">
+            <Label>Notification sound</Label>
+            <div className="flex items-center justify-between">
+              <span className="text-xs text-muted-foreground">
+                Play when agent needs attention
+              </span>
+              <Switch
+                id="sound-needs-attention"
+                checked={soundOnNeedsAttention}
+                onCheckedChange={(checked: boolean) => {
+                  setSoundOnNeedsAttention(checked);
+                  if (checked) {
+                    playSound(selectedSound);
+                  }
+                }}
+              />
+            </div>
+          </div>
+          {soundOnNeedsAttention && (
+            <div className="space-y-2">
+              <Label>Sound</Label>
               <DropdownMenu>
                 <DropdownMenuTrigger asChild>
                   <Button
@@ -398,7 +503,7 @@ export function SettingsPage({ onClose }: Props) {
                     size="sm"
                     className="w-full justify-between font-normal h-7 text-xs px-2"
                   >
-                    {agentType ? AGENT_LABEL[agentType] : "None"}
+                    {SOUNDS.find((s) => s.id === selectedSound)?.label ?? "Chime"}
                     <ChevronDown className="size-3 opacity-50" />
                   </Button>
                 </DropdownMenuTrigger>
@@ -407,231 +512,185 @@ export function SettingsPage({ onClose }: Props) {
                   className="w-[--radix-dropdown-menu-trigger-width]"
                 >
                   <DropdownMenuRadioGroup
-                    value={agentType}
-                    onValueChange={(v: string) => setAgentType(v as CodingAgentType | "")}
+                    value={selectedSound}
+                    onValueChange={(v: string) => {
+                      setSelectedSound(v as SoundId);
+                      playSound(v as SoundId);
+                    }}
                   >
-                    <DropdownMenuRadioItem value="">None</DropdownMenuRadioItem>
-                    {AGENT_TYPES.map((t) => (
-                      <DropdownMenuRadioItem key={t.value} value={t.value}>
-                        {t.label}
+                    {SOUNDS.map((s) => (
+                      <DropdownMenuRadioItem key={s.id} value={s.id}>
+                        {s.label}
                       </DropdownMenuRadioItem>
                     ))}
                   </DropdownMenuRadioGroup>
                 </DropdownMenuContent>
               </DropdownMenu>
               <p className="text-xs text-muted-foreground">
-                The coding agent used for background jobs.
+                Choose a sound to play when an agent transitions from working to needs attention.
               </p>
             </div>
-            {agentType && (
-              <div className="space-y-2">
-                <Label htmlFor="agent-command">
-                  Command <span className="text-muted-foreground font-normal">(optional)</span>
-                </Label>
-                <Input
-                  id="agent-command"
-                  placeholder="claude --dangerously-skip-permissions"
-                  value={agentCommand}
-                  onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                    setAgentCommand(e.target.value)
-                  }
-                />
-                <p className="text-xs text-muted-foreground">
-                  Custom command with arguments to run the agent. Leave empty to use the default
-                  command for the selected agent type.
-                </p>
-              </div>
-            )}
-          </div>
-        )}
+          )}
+        </div>
+      )}
 
-        {section === "defaults" && (
-          <div className="space-y-4 px-1">
-            <div className="space-y-2">
-              <div className="flex items-center justify-between">
-                <Label htmlFor="defaults-json">Default apps configuration</Label>
-                {!defaultsJson.trim() && (
-                  <Button
-                    type="button"
-                    variant="link"
-                    size="sm"
-                    className="h-auto p-0 text-xs"
-                    onClick={handleInsertTemplate}
-                  >
-                    Insert template
-                  </Button>
-                )}
-              </div>
-              <textarea
-                id="defaults-json"
-                className="w-full min-h-[160px] rounded-md border border-input bg-background px-2 py-1.5 text-xs font-mono ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 dark:bg-input/30"
-                placeholder='{"apps": [{"type": "vscode", ...}]}'
-                value={defaultsJson}
-                onChange={(e: React.ChangeEvent<HTMLTextAreaElement>) =>
-                  handleDefaultsChange(e.target.value)
-                }
-                autoCapitalize="off"
-                autoCorrect="off"
-                autoComplete="off"
-                spellCheck={false}
+      {activeSection === "web-server" && (
+        <div className="space-y-4 px-1">
+          <div className="space-y-2">
+            <Label htmlFor="web-server-port">Port</Label>
+            <Input
+              id="web-server-port"
+              type="number"
+              placeholder="3456 (default)"
+              value={webServerPort}
+              onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
+                setWebServerPort(e.target.value)
+              }
+              min={1}
+              max={65535}
+            />
+            <p className="text-xs text-muted-foreground">
+              Port the web server listens on for mobile access. Leave empty for the default (3456).
+              Requires restart.
+            </p>
+          </div>
+          <div className="space-y-2">
+            <div className="flex items-center justify-between">
+              <Label htmlFor="auto-start-tunnel">Auto-start tunnel</Label>
+              <Switch
+                id="auto-start-tunnel"
+                checked={autoStartTunnel}
+                onCheckedChange={setAutoStartTunnel}
               />
-              {defaultsError && <p className="text-xs text-destructive">{defaultsError}</p>}
-              <p className="text-xs text-muted-foreground">
-                Default apps configuration applied to Band worktrees that don't have a project-level{" "}
-                <code className="text-xs">.band/config.json</code>. Defines which apps (VS Code,
-                Cursor, Zed, iTerm, Chrome) to open and their layout. Leave empty to disable.
-              </p>
             </div>
+            <p className="text-xs text-muted-foreground">
+              Automatically start the web server and tunnel when the app launches.
+            </p>
           </div>
-        )}
+        </div>
+      )}
+    </>
+  );
 
-        {section === "notifications" && (
-          <div className="space-y-4 px-1">
-            <div className="space-y-2">
-              <Label>Notification sound</Label>
-              <div className="flex items-center justify-between">
-                <span className="text-xs text-muted-foreground">
-                  Play when agent needs attention
-                </span>
-                <Switch
-                  id="sound-needs-attention"
-                  checked={soundOnNeedsAttention}
-                  onCheckedChange={(checked: boolean) => {
-                    setSoundOnNeedsAttention(checked);
-                    if (checked) {
-                      playSound(selectedSound);
-                    }
-                  }}
-                />
-              </div>
-            </div>
-            {soundOnNeedsAttention && (
-              <div className="space-y-2">
-                <Label>Sound</Label>
-                <DropdownMenu>
-                  <DropdownMenuTrigger asChild>
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      className="w-full justify-between font-normal h-7 text-xs px-2"
-                    >
-                      {SOUNDS.find((s) => s.id === selectedSound)?.label ?? "Chime"}
-                      <ChevronDown className="size-3 opacity-50" />
-                    </Button>
-                  </DropdownMenuTrigger>
-                  <DropdownMenuContent
-                    align="start"
-                    className="w-[--radix-dropdown-menu-trigger-width]"
-                  >
-                    <DropdownMenuRadioGroup
-                      value={selectedSound}
-                      onValueChange={(v: string) => {
-                        setSelectedSound(v as SoundId);
-                        playSound(v as SoundId);
-                      }}
-                    >
-                      {SOUNDS.map((s) => (
-                        <DropdownMenuRadioItem key={s.id} value={s.id}>
-                          {s.label}
-                        </DropdownMenuRadioItem>
-                      ))}
-                    </DropdownMenuRadioGroup>
-                  </DropdownMenuContent>
-                </DropdownMenu>
-                <p className="text-xs text-muted-foreground">
-                  Choose a sound to play when an agent transitions from working to needs attention.
-                </p>
-              </div>
-            )}
-          </div>
-        )}
+  /* ── Menu items ─────────────────────────────────────────── */
 
-        {section === "web-server" && (
-          <div className="space-y-4 px-1">
-            <div className="space-y-2">
-              <Label htmlFor="web-server-port">Port</Label>
-              <Input
-                id="web-server-port"
-                type="number"
-                placeholder="3456 (default)"
-                value={webServerPort}
-                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                  setWebServerPort(e.target.value)
-                }
-                min={1}
-                max={65535}
-              />
-              <p className="text-xs text-muted-foreground">
-                Port the web server listens on for mobile access. Leave empty for the default
-                (3456). Requires restart.
-              </p>
-            </div>
-            <div className="space-y-2">
-              <div className="flex items-center justify-between">
-                <Label htmlFor="auto-start-tunnel">Auto-start tunnel</Label>
-                <Switch
-                  id="auto-start-tunnel"
-                  checked={autoStartTunnel}
-                  onCheckedChange={setAutoStartTunnel}
-                />
-              </div>
-              <p className="text-xs text-muted-foreground">
-                Automatically start the web server and tunnel when the app launches.
-              </p>
-            </div>
-          </div>
-        )}
-      </div>
-    );
-  }
+  const menuItems = (
+    <div className="flex flex-col gap-px">
+      <SettingsRow
+        label="Appearance"
+        value={themePreview}
+        active={activeSection === "appearance"}
+        onClick={() => setSection("appearance")}
+      />
+      <Separator />
+      <SettingsRow
+        label="General"
+        value={worktreesDirPreview}
+        active={activeSection === "general"}
+        onClick={() => setSection("general")}
+      />
+      <Separator />
+      <SettingsRow
+        label="Labels"
+        value={labelsPreview}
+        active={activeSection === "labels"}
+        onClick={() => setSection("labels")}
+      />
+      <Separator />
+      <SettingsRow
+        label="Coding Agent"
+        value={agentPreview}
+        active={activeSection === "coding-agent"}
+        onClick={() => setSection("coding-agent")}
+      />
+      <Separator />
+      <SettingsRow
+        label="Workspace Settings"
+        value={defaultsPreview}
+        active={activeSection === "defaults"}
+        onClick={() => setSection("defaults")}
+      />
+      <Separator />
+      <SettingsRow
+        label="Notifications"
+        value={notificationsPreview}
+        active={activeSection === "notifications"}
+        onClick={() => setSection("notifications")}
+      />
+      <Separator />
+      <SettingsRow
+        label="Web Server"
+        value={portPreview}
+        active={activeSection === "web-server"}
+        onClick={() => setSection("web-server")}
+      />
+    </div>
+  );
+
+  /* ── Layout ─────────────────────────────────────────────── */
 
   return (
-    <div>
-      <div className="flex items-center gap-1 mb-2 px-1">
-        <Button variant="ghost" size="icon-sm" onClick={onClose}>
-          <ChevronLeft className="size-5" />
-        </Button>
-        <h2 className="text-base font-semibold">Settings</h2>
+    <div className="flex flex-col lg:flex-row h-full">
+      {/* ── Left: menu panel ──────────────────────────────── */}
+      <div
+        className={`lg:w-72 lg:shrink-0 lg:border-r lg:border-border lg:block overflow-y-auto ${section !== "menu" ? "hidden" : ""}`}
+      >
+        {!hideTitle && (
+          <>
+            <div className="flex items-center gap-1 mb-2 px-1">
+              {onClose && (
+                <Button variant="ghost" size="icon-sm" onClick={onClose} className="lg:hidden">
+                  <ChevronLeft className="size-5" />
+                </Button>
+              )}
+              <h2 className="text-base font-semibold">Settings</h2>
+            </div>
+            <Separator />
+          </>
+        )}
+        {menuItems}
       </div>
-      <Separator />
-      <div className="flex flex-col gap-px">
-        <SettingsRow
-          label="Appearance"
-          value={themePreview}
-          onClick={() => setSection("appearance")}
-        />
-        <Separator />
-        <SettingsRow
-          label="General"
-          value={worktreesDirPreview}
-          onClick={() => setSection("general")}
-        />
-        <Separator />
-        <SettingsRow label="Labels" value={labelsPreview} onClick={() => setSection("labels")} />
-        <Separator />
-        <SettingsRow
-          label="Coding Agent"
-          value={agentPreview}
-          onClick={() => setSection("coding-agent")}
-        />
-        <Separator />
-        <SettingsRow
-          label="Workspace Settings"
-          value={defaultsPreview}
-          onClick={() => setSection("defaults")}
-        />
-        <Separator />
-        <SettingsRow
-          label="Notifications"
-          value={notificationsPreview}
-          onClick={() => setSection("notifications")}
-        />
-        <Separator />
-        <SettingsRow
-          label="Web Server"
-          value={portPreview}
-          onClick={() => setSection("web-server")}
-        />
+
+      {/* ── Right: detail panel ───────────────────────────── */}
+      <div
+        className={`flex-1 min-w-0 overflow-y-auto lg:block ${section === "menu" ? "hidden" : ""}`}
+      >
+        {activeSection && (
+          <div className="lg:px-4 lg:py-2">
+            <div className="flex items-center gap-1 mb-3 px-1">
+              <Button
+                variant="ghost"
+                size="icon-sm"
+                onClick={() => setSection("menu")}
+                className="lg:hidden"
+              >
+                <ChevronLeft className="size-5" />
+              </Button>
+              <h2 className="text-base font-semibold flex-1">
+                {SECTION_TITLES[activeSection]}
+              </h2>
+              <Tooltip>
+                <TooltipTrigger asChild>
+                  <Button
+                    variant="ghost"
+                    size="icon-sm"
+                    onClick={handleSave}
+                    disabled={activeSection === "defaults" && !!defaultsError}
+                    className="relative"
+                  >
+                    <Save className="size-5" />
+                    {isDirty && (
+                      <span className="absolute top-0.5 right-0.5 size-2 rounded-full bg-blue-500" />
+                    )}
+                  </Button>
+                </TooltipTrigger>
+                <TooltipContent>Save</TooltipContent>
+              </Tooltip>
+            </div>
+            <Separator className="mb-3" />
+            {sectionContent}
+          </div>
+        )}
       </div>
     </div>
   );
